requirements.txt
praw
requests
schedule
python-dotenv
groq